feat(core): improve resource read & write traits
This commit introduces two new buffer wrapper types to `deno_core`. The main
benefit of these new wrappers is that they can wrap a number of different
underlying buffer types. This allows for a more flexible read and write API on
resources that will require less copying of data between different buffer
representations.

- `BufView` is a read-only view onto a buffer. It can be backed by
  `ZeroCopyBuf`, `Vec<u8>`, and `bytes::Bytes`.
- `BufViewMut` is a read-write view onto a buffer. It can be cheaply
  converted into a `BufView`. It can be backed by `ZeroCopyBuf` or `Vec<u8>`.

Both new buffer views have a cursor. This means that the start point of the view
can be constrained to write / read from just a slice of the view. Only the start
point of the slice can be adjusted. The end point is fixed. To adjust the end
point, the underlying buffer needs to be truncated.

Readable resources have been changed to better cater to resources that do not
support BYOB reads. The basic `read` method now returns a `BufView` instead of
taking a `ZeroCopyBuf` to fill. This allows the operation to return buffers that
the resource has already allocated, instead of forcing the caller to allocate
the buffer. BYOB reads are still very useful for resources that support them,
so a new `read_byob` method has been added that takes a `BufViewMut` to fill.
`op_read` attempts to use `read_byob` if the resource supports it, which falls
back to `read` and performs an additional copy if it does not. For Rust->JS
reads this change should have no impact, but for Rust->Rust reads, this allows
the caller to avoid an additional copy in many scenarios. This combined with the
support for `BufView` to be backed by `bytes::Bytes` allows us to avoid one data
copy when piping from a `fetch` response into an `ext/http` response.

Writable resources have been changed to take a `BufView` instead of a
`ZeroCopyBuf` as an argument. This allows for less copying of data in certain
scenarios, as described above. Additionally a new `Resource::write_all` method
has been added that takes a `BufView` and continually attempts to write the
resource until the entire buffer has been written. Certain resources like files
can override this method to provide a more efficient `write_all` implementation.